# apps/api/src/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IdeiaERP Commerce Sync API",
    "version": "1.0.0",
    "description": "API REST para sincronização entre IdeiaERP e lojas virtuais"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "UUID",
        "description": "Token de autenticação obtido no endpoint /auth/login"
      }
    },
...
  },
...
  "paths": {
...
    "/api/v1/admin/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Lista todas as configurações",
        "description": "Retorna todas as configurações do sistema. Senhas são mascaradas.",
        "responses": {
          "200": {
            "description": "Lista de configurações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Setting"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Cria uma nova configuração",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "key",
                  "value"
                ],
                "properties": {
                  "key": {
                    "type": "string",
                    "example": "ERP_DB_HOST"
                  },
                  "value": {
                    "type": "string",
                    "example": "localhost"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuração criada"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/v1/admin/settings/{key}": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Busca uma configuração por chave",
        "description": "Retorna uma configuração específica. Senha é mascarada.",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Chave da configuração"
          }
        ],
        "responses": {
          "200": {
            "description": "Configuração encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Setting"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Configuração não encontrada"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Atualiza uma configuração existente",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "value"
                ],
                "properties": {
                  "value": {
                    "type": "string",
                    "example": "localhost"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuração atualizada"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Configuração não encontrada"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Remove uma configuração",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuração removida"
          },
          "404": {
            "description": "Configuração não encontrada"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    },
    "/api/v1/admin/settings/erp/test-connection": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Testa a conexão com o ERP-DB",
        "responses": {
          "200": {
            "description": "Conexão bem-sucedida"
          },
          "503": {
            "description": "Não foi possível conectar"
          }
        }
      }
    },
...
  }
}